requirements.txt
alpaca-py==0.31.0
annotated-types==0.7.0
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
bottle==0.13.2
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dateparser==1.2.0
debugpy==1.8.7
decorator==5.1.1
exceptiongroup==1.2.2
exchange_calendars==4.5.6
executing==2.1.0
fonttools==4.54.1
humanize==4.11.0
idna==3.10
imageio==2.35.1
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
korean-lunar-calendar==0.3.1
lightweight_charts @ git+https://github.com/drew2323/lightweight-charts-python.git@b3368578320b811f25df3d664ae1bb53069c31cd
llvmlite==0.39.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
msgpack==1.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numba==0.56.4
numpy==1.23.5
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pandas_market_calendars==4.4.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
proxy-tools==0.1.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyluach==2.2.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyobjc-framework-Security==10.3.1
pyobjc-framework-WebKit==10.3.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywebview==5.3.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
schedule==1.2.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sseclient-py==1.8.0
stack-data==0.6.3
tenacity==9.0.0
threadpoolctl==3.5.0
tomli==2.0.2
toolz==1.0.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
vectorbtpro
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
widgetsnbextension==4.0.13
